[CONFIGURATION] * BUG: Automatically search libraries in lib64 before searching in lib. * BUG: Fixed the $FFI_LIB test in main/configure.ac.
[DEVELOPMENT ENVIRONMENT]
* BUG: Tips of the days are displayed correctly again.

[INTERPRETER]
* BUG: Val() now takes the thousand separator into account when this 
  separator is a space character.
* BUG: Removed the 64 bits warning.
* BUG: Fixes the 'lib64' problem. Now components are searched in the 'lib64' 
  directory on 64 bits OS. If they are not stored in 'lib64', then the 
  default 'lib' directory is used.
* NEW: Passing argument by reference is not allowed anymore if not 
  explicitely specified in the method declaration.

[COMPILER]
* NEW: Passing argument by reference is not allowed anymore if not 
  explicitely specified in the method declaration.

[INFORMER]
* BUG: Fixes the 'lib64' problem. Now components are searched in the 'lib64' 
  directory on 64 bits OS. If they are not stored in 'lib64', then the 
  default 'lib' directory is used.


git-svn-id: svn://localhost/gambas/trunk@1108 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gbc_trans_common.h
@@ -76,7 +76,7 @@ typedef
     TRANS_PARAM param[MAX_PARAM_FUNC];
     PATTERN *start;
     int line;
-    uint byref;
+    uint64_t byref;
     unsigned vararg : 1;
     }
   PACKED
